Prepared directory structure for navigation
