+++ app/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Provider } from 'react-redux';
 import { createLogger } from 'redux-logger';
-import AppContainer from './containers/AppContainer';
+import Home from './scenes/home';
 import configureStore from './store';
 
 const store = configureStore();
@@ -11,7 +11,7 @@ class App extends Component {
   render () {
     return (
       <Provider store={store}>
-        <AppContainer />
+        <Home />
       </Provider>
     );
   }
