+++ app/scenes/home/styles.js
@@ -1,0 +1,8 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  scene: {
+    marginTop: 10,
+    flex: 1
+  }
+});
